better link mailhog installation with mailhog server parameters + fix postgresql storage class

+++ mailhog/uninstall_mailhog.sh
@@ -1,7 +1,13 @@
 #! /bin/sh
 
-NAMESPACE=mail
+if [ "$#" -ne "2" ] ; then
+  echo "Usage: $0 <namespace> <release-name>"
+  exit 1
+fi
 
-helm uninstall --namespace ${NAMESPACE} mailhog
+NAMESPACE=$1
+RELEASE_NAME=$2
+
+helm uninstall --namespace ${NAMESPACE} ${RELEASE_NAME}
 
 kubectl delete namespace ${NAMESPACE}
